backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Abhishek Menon",email = "128571215+Nasper-github@users.noreply.github.com"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "lxml (>=6.0.2,<7.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]
